# docs/api/openapi-auth.yaml
openapi: 3.0.0
info:
  title: MCP-RS 認証API
  description: |
    PostgreSQL、Redis、JWTを統合した認証API
    - ユーザー登録・ログイン
    - トークンベース認証
    - セッション管理
  version: 0.15.0
  contact:
    name: MCP-RS Team
    url: https://github.com/n-takatsu/mcp-rs
  license:
    name: MIT OR Apache-2.0
    url: https://github.com/n-takatsu/mcp-rs/blob/main/LICENSE-MIT

servers:
  - url: http://localhost:3000
    description: ローカル開発サーバー
  - url: https://api.example.com
    description: 本番環境サーバー

tags:
  - name: Authentication
    description: 認証エンドポイント

paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: ユーザー登録
      description: 新しいユーザーアカウントを作成します。パスワードは自動的にArgon2でハッシュ化されます。
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              basic:
                summary: 基本的な登録リクエスト
                value:
                  username: alice
                  password: SecurePass123!
                  email: alice@example.com
      responses:
        '201':
          description: ユーザー登録成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
              examples:
                success:
                  value:
                    user_id: 550e8400-e29b-41d4-a716-446655440000
                    username: alice
                    email: alice@example.com
                    message: User registered successfully
        '400':
          description: 不正なリクエスト（弱いパスワード、無効なメールなど）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                weakPassword:
                  value:
                    error: WeakPassword
                    message: Password must be at least 8 characters and contain uppercase, lowercase, numbers, and special characters
        '409':
          description: ユーザー名またはメールが既に存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                duplicate:
                  value:
                    error: UserAlreadyExists
                    message: User with this email already exists

  /auth/login:
    post:
      tags:
        - Authentication
      summary: ログイン
      description: |
        ユーザー認証を行い、JWTトークンペアを発行します。
        Redis有効時はセッションも作成されます。
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              basic:
                summary: 基本的なログイン
                value:
                  email: alice@example.com
                  password: SecurePass123!
                  remember_me: false
              rememberMe:
                summary: 長期セッション
                value:
                  email: alice@example.com
                  password: SecurePass123!
                  remember_me: true
      responses:
        '200':
          description: ログイン成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              examples:
                success:
                  value:
                    access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    refresh_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    token_type: Bearer
                    expires_in: 3600
                    user:
                      id: 550e8400-e29b-41d4-a716-446655440000
                      username: alice
                      email: alice@example.com
                      roles:
                        - User
        '401':
          description: 認証失敗（メールまたはパスワードが正しくない）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidCredentials:
                  value:
                    error: InvalidCredentials
                    message: Invalid email or password

  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: トークンリフレッシュ
      description: リフレッシュトークンを使用して新しいトークンペアを取得します。
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
            examples:
              basic:
                value:
                  refresh_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: トークンリフレッシュ成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
              examples:
                success:
                  value:
                    access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    refresh_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    token_type: Bearer
                    expires_in: 3600
        '401':
          description: 無効なリフレッシュトークン
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidToken:
                  value:
                    error: InvalidToken
                    message: Invalid or expired refresh token

  /auth/logout:
    post:
      tags:
        - Authentication
      summary: ログアウト
      description: ユーザーをログアウトさせます（将来的にトークンのブラックリスト化を実装予定）。
      operationId: logoutUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: ログアウト成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              examples:
                success:
                  value:
                    message: Logged out successfully

  /auth/me:
    get:
      tags:
        - Authentication
      summary: 現在のユーザー情報取得
      description: 認証済みユーザーの情報を取得します（未実装）。
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: ユーザー情報取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
              examples:
                success:
                  value:
                    id: 550e8400-e29b-41d4-a716-446655440000
                    username: alice
                    email: alice@example.com
                    roles:
                      - User
        '401':
          description: 認証が必要
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '501':
          description: 未実装
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                notImplemented:
                  value:
                    error: NotImplemented
                    message: Use AuthMiddleware to extract user

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWTアクセストークン

  schemas:
    RegisterRequest:
      type: object
      required:
        - username
        - password
        - email
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          description: ユーザー名
          example: alice
        password:
          type: string
          minLength: 8
          maxLength: 128
          format: password
          description: |
            パスワード（要件: 8文字以上、大文字・小文字・数字・特殊文字を含む）
          example: SecurePass123!
        email:
          type: string
          format: email
          description: メールアドレス（一意）
          example: alice@example.com

    RegisterResponse:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
          description: 生成されたユーザーID
          example: 550e8400-e29b-41d4-a716-446655440000
        username:
          type: string
          description: ユーザー名
          example: alice
        email:
          type: string
          format: email
          description: メールアドレス
          example: alice@example.com
        message:
          type: string
          description: 成功メッセージ
          example: User registered successfully

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: メールアドレス
          example: alice@example.com
        password:
          type: string
          format: password
          description: パスワード
          example: SecurePass123!
        remember_me:
          type: boolean
          default: false
          description: 長期セッション有効化（30日）
          example: false

    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWTアクセストークン（短期間有効）
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refresh_token:
          type: string
          description: JWTリフレッシュトークン（長期間有効）
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        token_type:
          type: string
          enum:
            - Bearer
          description: トークンタイプ
          example: Bearer
        expires_in:
          type: integer
          format: int64
          description: アクセストークンの有効期限（秒）
          example: 3600
        user:
          $ref: '#/components/schemas/UserInfo'

    RefreshRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          description: リフレッシュトークン
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...

    RefreshResponse:
      type: object
      properties:
        access_token:
          type: string
          description: 新しいJWTアクセストークン
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refresh_token:
          type: string
          description: 新しいJWTリフレッシュトークン
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        token_type:
          type: string
          enum:
            - Bearer
          description: トークンタイプ
          example: Bearer
        expires_in:
          type: integer
          format: int64
          description: アクセストークンの有効期限（秒）
          example: 3600

    UserInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ユーザーID
          example: 550e8400-e29b-41d4-a716-446655440000
        username:
          type: string
          description: ユーザー名
          example: alice
        email:
          type: string
          format: email
          nullable: true
          description: メールアドレス
          example: alice@example.com
        roles:
          type: array
          items:
            type: string
          description: ユーザーロール
          example:
            - User

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: エラータイプ
          example: InvalidCredentials
        message:
          type: string
          description: エラーメッセージ
          example: Invalid email or password

    SuccessResponse:
      type: object
      properties:
        message:
          type: string
          description: 成功メッセージ
          example: Logged out successfully